+++ tests/testthat/test-tree_rearrange.R
@@ -356,6 +356,12 @@ test_that("CollapseNode() handles node labels", {
                AddTip(bal6, where = 1)[["node.label"]])
 })
 
+test_that("MakeTreeBinary() edge lengths", {
+  tree <- ape::read.tree(text = "(a:1, b:2, c:3, d:4);")
+  # Not yet supported; ensure they are removed
+  expect_null(MakeTreeBinary(tree)$edge.length)
+})
+
 test_that("Binarification is uniform", {
   set.seed(0)
   Test <- function(tree, nTree, nSamples = 200L, ape = FALSE) {